+++ ChefMaster/Assets/Scripts/Customer.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+
+using System;
 using UnityEngine;
 using UnityEngine.UI;
 using System.Linq;
@@ -18,21 +20,37 @@ using System.Linq;
 
 public class Customer : MonoBehaviour,InterfaceFunctions
 {
+    public static event Action<Player> IncreaseScore;
+
     public SaladClass currentList;
 
     public List<InventoryEnum> dummyList;
+
+    public Image waitTimeIndicator;
+    public float customerWaitTime;
+    public float currentWaitTime;
+    public float bufferTime;
+
+    public Transform[] orderIndicator;
+    public List<Vegetable> currentOrder;
+
+    public float speedMultiplier;
+    public float timeValue;
+    GameManager gameManager;
+
+    public bool isCustomerActive;
 
     public bool OnItemDrop(PlayerController playerScript)
     {
         if (playerScript.gotChoppedItem)
         {
-
             playerScript.gotChoppedItem = false;
-
             Destroy(playerScript.bowl);
-
             if (currentList.vegetables.Length == playerScript.combinationList.Count)
             {
+
+
+
                 /*   eaqualItems = saladClass.vegetables.Intersect(combinationList);
                   var repeatValue = saladClass.vegetables.Distinct();
                   if(eaqualItems.Count() == saladClass.vegetables.Length-repeatValue.Count())
@@ -48,9 +66,7 @@ public class Customer : MonoBehaviour,InterfaceFunctions
                   }*/
 
                 dummyList = new List<InventoryEnum>(playerScript.combinationList);
-
                 for (int i = 0; i < currentList.vegetables.Length; i++)
-
                 {
                     for (int j = 0; j < dummyList.Count; j++)
                     {
@@ -60,20 +76,37 @@ public class Customer : MonoBehaviour,InterfaceFunctions
                             dummyList.Remove(dummyList[j]);
                             break;
                         }
-                        
                     }
-
                 }
-                
             }
             else
             {
                 playerScript.correctItem = false;
             }
-           
+
+            if (playerScript.correctItem)
+            {
+                ResetCustomer();
+                playerScript.score+=5;
+
+                if (IncreaseScore != null)
+                    IncreaseScore(playerScript.playerContorl);
+
+
+                //reward Player
+                Debug.Log("Reward Player : " +playerScript.score );
+
+                if (timeValue >= 0.7f)
+                {
+                    Debug.Log("playergets pickup");
+                }
+            }
+            else
+            {
+                //player penalty
+                Debug.Log("Customer is angry");
+            }
         }
-
-
         playerScript.combinationList = new List<InventoryEnum>();
         return true;
     }
@@ -83,20 +116,85 @@ public class Customer : MonoBehaviour,InterfaceFunctions
         return false;
     }
 
+    private void Awake()
+    {
+        gameManager = GameManager.Instance;
+        currentOrder = new List<Vegetable>();
+    }
 
     // Start is called before the first frame update
     void Start()
     {
-        
+       
     }
 
     // Update is called once per frame
     void Update()
     {
-        currentList = GameManager.Instance.saladList[1];
+        if (isCustomerActive)
+        {
+            if (currentWaitTime <= 0 )
+            {
+                ResetCustomer();
+
+            }
+            else
+            {
+                currentWaitTime -= Time.deltaTime* speedMultiplier;
+              timeValue = currentWaitTime / customerWaitTime;
+                waitTimeIndicator.fillAmount =  currentWaitTime / customerWaitTime ;
+            }
+        }
+    }
+    private void OnEnable()
+    {
+        GetSaladList();
+    }
+
+    private void OnDisable()
+    {
+       
+    }
+    void GetSaladList()
+    {
+        waitTimeIndicator.fillAmount = 1f;
+        var value = UnityEngine.Random.Range(0, gameManager.saladList.Length);
+        currentList = gameManager.saladList[value];
+
+        customerWaitTime = 0f;
+        for(int i = 0; i< currentList.vegetables.Length; i++)
+        {
+           var currentObject= Instantiate( gameManager.vegetablesPrefab[(int)currentList.vegetables[i]],orderIndicator[i]);
+            currentObject.transform.localPosition = Vector3.zero;
+            currentOrder.Add(currentObject);
+
+            customerWaitTime += currentObject.choppingTime;
+        }
+        customerWaitTime = (customerWaitTime*2)+ bufferTime;
+        currentWaitTime = customerWaitTime;
+        isCustomerActive = true;
+    }
+
+    void ResetCustomer()
+    {
+
+        isCustomerActive = false;
+        for (int i = 0; i < currentOrder.Count; i++)
+        {
+
+            Destroy(currentOrder[i].gameObject);
+        }
+        currentOrder = new List<Vegetable>();
+        waitTimeIndicator.fillAmount = 0f;
+        StartCoroutine(SpawnNewCustomer());
     }
 
 
+    IEnumerator SpawnNewCustomer()
+    {
+        yield return new WaitForSeconds(3f);
+        GetSaladList();
 
+    }
 
 }