+++ ChefMaster/Assets/Scripts/GameManager.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 /// <summary>
 /// 
@@ -30,7 +31,11 @@ public class GameManager : MonoBehaviour
 
     public SaladClass[] saladList;
 
-    public GameObject[] vegetables;
+    public Sprite[] vegetables;
+    public Vegetable[] vegetablesPrefab;
+
+    public PlayerController player_P1;
+    public PlayerController player_P2;
 
     // Start is called before the first frame update
     void Start()